factorize bench for solver cartesian struct
